move LicenseCollector to new library
